ai_learning_platform/requirements.txt
Django>=5.0.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-environ==0.11.2
django-cors-headers==4.3.1
psycopg2-binary>=2.9.10
celery>=5.4.0
redis>=5.0.0
django-celery-beat>=2.6.0
openai==1.30.1
drf-spectacular>=0.27.0
pytest>=8.0.0
pytest-django>=4.8.0
factory-boy>=3.3.0
Faker>=25.0.0
coverage>=7.5.0
gunicorn>=22.0.0
structlog>=24.1.0
setuptools
django-lifecycle
